Change method definition location
