+++ control/Field.java
@@ -1,6 +1,6 @@
 package control;
 
-import java.util.Random;
+import java.util.*;
 import models.*;
 import options.UnitType;
 
@@ -9,13 +9,52 @@ public class Field extends Information {
 
     private static Random rand = new Random();
 
-    // 本体に影響がないようにfieldmapのディープコピーを渡すメソッド
+    // 周囲rangeマスのfield範囲内の座標一覧を{{y, x}, {{y, x}, ...}の配列にして返す
+    public static int[][] surroundField(int y, int x, int range) {
+        ArrayList<int[]> surround = new ArrayList<>();
+
+        for(int i = y - range; i <= y + range; i++) {
+            for(int j = x - range; j <= x + range; j++) {
+                if(isOutOfField(i, j))
+                    continue;
+                int[] yx = {i, j};
+                surround.add(yx);
+            }
+        }
+
+        int[][] res = new int[surround.size()][2];
+        for(int i = 0; i < surround.size(); i++)
+            res[i] = surround.get(i);
+
+        return res;
+    }
+
+    public static int[][] surroundField(int y, int x) {
+        return surroundField(y, x, 1);
+    }
+
+    public static boolean isOutOfField(int y, int x) {
+        return (x < 0 || x >= Field.MAX_X || y < 0 || y >= Field.MAX_Y);
+    }
+
     public static Cell[][] getClone() {
         Cell[][] res = new Cell[MAX_Y][MAX_X];
         for(int i = 0; i < MAX_Y; i++)
             for(int j = 0; j < MAX_X; j++)
                 res[i][j] = fieldmap[i][j].clone();
         return res;
+    }
+
+    public static void updateAvailable(int y, int x, boolean available) {
+        int range = AVAILABLE_RANGE + (available ? 0 : 1);
+        fieldmap[y][x].available = available;
+
+        int[][] surround = surroundField(y, x, range);
+        for(int i = 0; i < surround.length; i++) {
+            Cell cell = fieldmap[surround[i][0]][surround[i][1]];
+            if(!(cell instanceof Unit))
+                cell.available = available;
+        }
     }
 
     // fieldの初期化
@@ -48,7 +87,6 @@ public class Field extends Information {
         }
     }
 
-    // 地雷の設定
     private static void setMines() {
         int count = 0;
         while(count < minesCount) {
