+++ models/Unit.java
@@ -19,7 +19,6 @@ public class Unit extends Cell {
         detected  = true;
     }
 
-    // ユニット移動 (前にいたところは平地になる)
     public void moveTo(Cell dest) {
         Information.fieldmap[y][x] = new Flatland(y, x, surroundMines, detected);
 
@@ -30,7 +29,6 @@ public class Unit extends Cell {
         Information.fieldmap[dest.y][dest.x] = this;
     }
 
-    // ユニットが死んだとき
     public void death() {
         Information.fieldmap[y][x] = new Flatland(y, x, surroundMines - 1, detected);
         dead = true;
@@ -43,17 +41,5 @@ public class Unit extends Cell {
 
     public boolean isAlly() {
         return type.isAlly();
-    }
-
-    public void updateAvailable(boolean available) {
-        int range = Information.AVAILABLE_RANGE + (available ? 0 : 1);
-        this.available = available;
-
-        int[][] surround = Information.surroundField(y, x, range);
-        for(int i = 0; i < surround.length; i++) {
-            Cell cell = Information.fieldmap[surround[i][0]][surround[i][1]];
-            if(!(cell instanceof Unit))
-                cell.available = available;
-        }
     }
 }
