+++ ui/animations/ExplodeAnimation.java
@@ -1,8 +1,8 @@
 package ui.animations;
 
-import control.*;
-import models.*;
 import java.util.*;
+import control.Field;
+import models.*;
 
 // 爆発アニメーション
 public class ExplodeAnimation extends Animation {
@@ -41,7 +41,7 @@ public class ExplodeAnimation extends Animation {
     private void explode(int y, int x) {
         fieldmap[y][x].character = "*";
 
-        int[][] surround = Information.surroundField(y, x);
+        int[][] surround = Field.surroundField(y, x);
         for(int i = 0; i < surround.length; i++) {
             int y2 = surround[i][0];
             int x2 = surround[i][1];
